+++ CHANGELOG.md
@@ -11,6 +11,9 @@ The main work (all changes without a GitHub username in brackets in the below li
 
 ## __WORK IN PROGRESS__
 
+- matter.js
+    - Fix: Fixes crash when decommissioning a node while a reconnect is in progress (#3237)
+
 - @matter/protocol
     - Adjustment: OTA update files are now stored per software version, allowing different updates to be served to different nodes simultaneously. Former files are migrated.
     - Fix: Ensures that production certificates are always stored correctly